refactor: 댓글 관련 컴포넌트 분리
