+++ src/app/questions/[questionid]/answers/[answerid]/page.tsx
@@ -1,9 +1,7 @@
 "use client";
-import { useState } from "react";
-import { Button } from "@/components/ui/button";
-import { Pagination } from "@/components/ui/pagination";
-import { TextArea } from "@/components/ui/textArea";
-import AnswerContent from "./components/AnswerContent";
+
+import CommentSection from "./components/CommentSection";
+import AnswerContents from "./components/AnswerContents";
 
 export default function AnswerDetailPage() {
   // Mock data for comments
@@ -17,83 +15,18 @@ export default function AnswerDetailPage() {
       date: `2025.03.${(i % 30) + 1}`,
     }));
 
-  const [currentPage, setCurrentPage] = useState(1);
-  const itemsPerPage = 10;
-
-  const handleMovePage = (page: number) => {
-    setCurrentPage(page);
-  };
-
-  const currentComments = comments.slice(
-    (currentPage - 1) * itemsPerPage,
-    currentPage * itemsPerPage
-  );
-
   return (
     <div className="container mx-auto pt-[40px] md:px-6">
       <div className="pb-[var(--space-24)] txt-2xl-b">답변 상세보기</div>
 
-      <AnswerContent />
+      <AnswerContents />
 
       {/* 본문과 텍스트 사이의 패딩 */}
       <div className="mb-[150px]"></div>
 
-      {/* 댓글 입력 영역 */}
-      <div className="mt-[40px] mb-[16px]">
-        <label className="txt-sm-b" htmlFor="comment">
-          <span className="text-[var(--gray-02)]">댓글 수 </span>
-          <span className="text-black">{comments.length}</span>
-        </label>
-        <div className="mb-[10px]" />
+     
 
-        <div className="flex gap-2">
-          {/* 좌측 박스: 작성자 + 입력창 */}
-          <TextArea placeholder="Type your message here" />
-          <Button className="w-[80px] h-[120px]" variant="outline">
-            등록
-          </Button>
-        </div>
-      </div>
-
-      {/* 본문과 텍스트 사이의 패딩 */}
-      <div className="mb-[58px]"></div>
-
-      {/* 댓글 리스트 */}
-      <div className="mt-[30px] space-y-8">
-        {currentComments.map((comment) => (
-          <div key={comment.id} className="border-b pb-4 flex gap-4">
-            {/* 프로필 이미지 */}
-            <span className="w-[36px] h-[36px] inline-block bg-[var(--gray-01)] rounded-full shrink-0"></span>
-
-            {/* 댓글 내용 */}
-            <div className="flex-1">
-              <div className="text-sm font-bold mb-1">{comment.author}</div>
-              <div className="text-base mb-1">{comment.content}</div>
-              <div className="text-xs text-[var(--gray-02)]">{comment.date}</div>
-            </div>
-
-            {/* 수정/삭제 버튼 */}
-            <div className="flex flex-col items-end gap-2">
-              <Button variant="gray" size="sm">
-                수정
-              </Button>
-              <Button variant="gray" size="sm">
-                삭제
-              </Button>
-            </div>
-          </div>
-        ))}
-      </div>
-
-      {/* 페이지네이션 */}
-      <Pagination
-        totalCount={comments.length}
-        itemsPerPage={itemsPerPage}
-        pageNumber={currentPage}
-        currentPageBlock={currentPage}
-        handleMovePageBlock={() => {}}
-        handleMovePage={handleMovePage}
-      />
+      <CommentSection comments={comments} />
     </div>
   );
 }
